feat: move main classes to \src
