+++ main.py
@@ -1,4 +1,4 @@
-import controller
+from src import controller
 
 if __name__ == '__main__':
     c = controller.Controller()
